Eliminar clase collection.Card, ahora utiliza game.Card
